requirements.txt
streamlit
torch
torchvision
transformers
diffusers
matplotlib
pillow
numpy